Add upgrade effects
Products and the mouse have a base biscuits per second/click along
with a bonus and multiplier. These can be affected by upgrades.
When an upgrade is first purchased, its effects are applied. I also
plan to apply all the effects of purchased upgrades when a saved
game state is loaded.